modules/location-consumer/requirements.txt
psycopg2-binary==2.9.1
grpcio==1.41.0
grpcio-tools==1.41.0
kafka-python==2.0.2
